feat provider: Support keycloak user role mapping resource
Adds a resource type to map Keycloak roles to users (either normal users or service accounts).

This resource type does not currently support imports because the only available information during importing is the role ID, however the user ID is also required.

This can theoretically be fixed by specifying a custom ID format and modifying it appropriately, however that is a bit messy and I'm not yet sure how to deal with it.