fix provider/realm: Work around SMTP password issue
SMTP passwords are not returned in plaintext by Keycloak (which is principally good!),
but this means that Terraform will constantly detect a password change if the SMTP
password field is set as Keycloak literally returns a string of just asterisks.

This will ignore changes to that field in this case, which makes it possible to create
an initial SMTP config but not to update the password in an existing one.

This issue will be addressed in a future commit.