feat build: Add Repeatr formula for reproducible builds
Adds a Repeatr[1] formula for repeatable and reproducible builds. This
should always result in byte-for-byte identical binaries.

The only dependency (terraform) is pinned to its v0.10.0 tag in this
commit.

Additionally a build script is added that sets up a release folder
structure similar to the structure I use in kontemplate[2].

This relates to issue #17

[1]: http://repeatr.io/
[2]: https://github.com/tazjin/kontemplate