chore: remove Repeatr formula and references in README.md
Because Repeatr is really painful to setup and there are other alternatives
to ensuring repeatable builds nowadays, any hints of Repeatr usage is
removed.

There's no doubt we need repeatable builds rather than going back to the
cowboy style of using the latest-off-of-master versions dependencies
and whatever golang version. That'll come in a later PR.