refactor: use idiomatic URL styled imports throughout the project
Previously, imports of internal sub packages was not formed as an URL
as is the idiomatic go way of referencing packages outside golang's
standard library.

It worked as expected by enforcing a particular `$GOPATH` setup. The
downside having such a custom setup, is it makes it more difficult for
developers getting everything working -- both for building from source
and making IDEs happy. As a result it raises the bar for newcomers
easily contributing to the project.