refactor: simply Makefile as go modules always ensures dependencies are in place
Up until now, we've had an explicit step to get hold of dependencies
that needed to run before trying to test / build etc.

Go modules is hooked into all the different subcommands of the `go` CLI,
ensuring that dependencies are in place before executing the said subcommand.
In practise that means that whenever we run `$ go build`, the `go` CLI
will firstly ensure the required dependencies are in place, before actually
trying to build the project.