fix(ci): run go vet before go fmt
Because absolutely all subcommands of the `go` CLI might end up installing
dependencies and therefore causing output to be written to stdout,
we move the `go vet` command *before* `go fmt`. That's because we break
the build if `go fmt` causes any output to stdout, which it does when it's
the first `go` run.