feat(ci): make Travis CI upload release binaries upon new tags
These changes makes use of Travis CI to build binaries for linux, darwin,
freebsd and windows, and upload those to github.com whenever new git tags
are pushed.

The intention of doing is to lower the barrier for cutting new releases
and ensuring the resulting binaries aren't dependant on how we as developers
has setup our machines.

In practise this means that when we want to publish a new release of this
terraform provider, after all commits has been landed in the master branch,
a new git tag has to be created and pushed as well:

```
$ git tag v1.0.0
$ git push origin --tags
```

That should make Travis CI build the project as usual, and finish off
by uploading the successfully built binaries to github.com, making them
visible as an "asset" on the [Releases page](https://github.com/tazjin/terraform-provider-keycloak/releases).